refactor(calling): address PR review comments
- Restructure playwright/tests/calling/ → playwright/calling/{tests,utils}
- Rename global.setup.ts → oauth.setup.ts
- Rename fullSetup → initAndRegister
- Extract ServiceIndicator type to constants
- Split callingUtils into setup.ts and registration.ts
- Add inline comments for Chrome launch flags in playwright.config.ts
